upstream(iota-execution): Pull IotaDataStore into a separate module
