+++ iota-execution/latest/iota-adapter/src/programmable_transactions/mod.rs
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod context;
+pub mod data_store;
 pub mod execution;
 pub mod linkage_view;
 pub mod package_metadata;
